multiple fixes

+++ src/transformers/utils/generic.py
@@ -783,9 +783,9 @@ class TransformersKwargs(TypedDict, total=False):
             Turn this on to return the intermediary attention scores.
         output_router_logits (`Optional[bool]`, *optional*):
             For MoE models, this allows returning the router logits to compute the loss.
-        cu_seq_lens_q (`torch.LongTensor`, *optional*)
+        cu_seq_lens_q (`torch.IntTensor`, *optional*)
             Gets cumulative sequence length for query state.
-        cu_seq_lens_k (`torch.LongTensor`, *optional*)
+        cu_seq_lens_k (`torch.IntTensor`, *optional*)
             Gets cumulative sequence length for key state.
         max_length_q (`int`, *optional*):
             Maximum sequence length for query state.
@@ -801,8 +801,8 @@ class TransformersKwargs(TypedDict, total=False):
     output_hidden_states: bool | None
     output_attentions: bool | None
     output_router_logits: bool | None
-    cu_seq_lens_q: torch.LongTensor | None
-    cu_seq_lens_k: torch.LongTensor | None
+    cu_seq_lens_q: torch.IntTensor | None
+    cu_seq_lens_k: torch.IntTensor | None
     max_length_q: int | None
     max_length_k: int | None
     position_ids: torch.LongTensor | None
